Search with title or status
