+++ frontend/src/App.js
@@ -1,13 +1,11 @@
+//React imports
 import { Routes, Route, Navigate, useLocation } from "react-router-dom";
 import { useState, useEffect } from "react";
 import "./App.css";
 
+//LoggedIn
 import NavbarLoggedIn from "./Components/Common/NavbarLoggedIn";
-import NavbarLoggedOut from "./Components/Common/NavbarLoggedOut";
-import Footer from "./Components/Common/Footer";
-
 import HomeLoggedIn from "./Pages/HomeLoggedIn";
-import HomeLoggedOut from "./Pages/HomeLoggedOut";
 import Categories from "./Pages/Categories";
 import Profile from "./Pages/Profile";
 import AboutUs from "./Pages/AboutUs";
@@ -18,30 +16,40 @@ import UserPolicy from "./Pages/UserPolicy";
 import Terms from "./Pages/Terms";
 import PrivacyPolicy from "./Pages/PrivacyPolicy";
 import Dashboard from "./Pages/Dashboard";
-import Auth from "./Pages/Auth";
 import AccountSettings from "./Pages/AccountSettings";
 import MoreSettings from "./Pages/MoreSettings";
 import Reminders from "./Pages/Reminders";
 import Error404 from "./Pages/Error404";
-
 import UpdateProfile from "./Components/Profile/UpdateProfile";
-
-import MyCalendar from "./Components/Calendar/Calendar";
-import MyCalendar2 from "./Components/Calendar/Calendar2"
+import MyCalendar from "./Components/Calendar/Calendar"
 import AddEvents from "./Components/Calendar/AddEvents";
 import UpdateEvent from "./Components/Calendar/UpdateEvent";
-
-import LightMode from "./Logic/LightMode";
-import DarkMode from "./Logic/DarkMode";
 import UploadEvents from "./Components/Calendar/UploadEvents";
 
+//LoggedOut
+import NavbarLoggedOut from "./Components/Common/NavbarLoggedOut";
+import HomeLoggedOut from "./Pages/HomeLoggedOut";
+import Auth from "./Pages/Auth";
+
+//Universal
+import Footer from "./Components/Common/Footer";
+import LightMode from "./Logic/LightMode";
+import DarkMode from "./Logic/DarkMode";
+
+
+
 function App() {
+
+    //Hooks
+    const location = useLocation();
+    
+    //States
     const user = localStorage.getItem("token");
     const auth = localStorage.getItem("auth");
-
     const [mode, setMode] = useState(localStorage.getItem("mode") || "light");
-
     const [eventsCount, setEventsCount] = useState(0);
+
+    //Handle functions
     const handleEventsCountChange = (count) => {
         setEventsCount(count);
     };
@@ -62,8 +70,6 @@ function App() {
             setMode("light");
         }
     };
-
-    const location = useLocation();
 
     const pathsWithoutNavbarFooter = ["/auth"];
     const shouldRenderNavbarFooter = !pathsWithoutNavbarFooter.includes(
@@ -203,12 +209,8 @@ function App() {
                                     element={<PrivacyPolicy mode={mode} />}
                                 />
                                 <Route
-                                    path="/events"
-                                    element={<MyCalendar mode={mode} />}
-                                />
-                                <Route
                                     path="/events2"
-                                    element={<MyCalendar2 mode={mode} />}
+                                    element={<MyCalendar mode={mode} />}
                                 />
                                 <Route
                                     path="/event/:id/update"
